+++ llm_invoker.py
@@ -1,16 +1,19 @@
 from chain_manager import ChainManager
 
 class LLMInvoker:
-    def __init__(self, query, config):
-        self.query = query
+    def __init__(self, questions, config):
+        self.questions = questions
         self.chain_manager = ChainManager(config['llm_to_be_evaluated'], config['embedding'], config['vectorstore'])
-        self.qa_chain = self.chain_manager.create_chain()
+        self.chain = self.chain_manager.create_chain()
 
     def invoke(self):
-        result = self.qa_chain.invoke(input={"input": self.query})
-        new_result = {
-            "query": result["input"],
-            "result": result["answer"],
-            "source_documents": result["context"],
-        }
-        return new_result
+        results = []
+        for question in self.questions:
+            result = self.chain.invoke(input={"input": question})
+            new_result = {
+                "query": result["input"],
+                "result": result["answer"],
+                #"source_documents": result["context"],
+            }
+            results.append(new_result)
+        return results
