+++ main.py
@@ -3,9 +3,16 @@ from config_loader import ConfigLoader
 from llm_invoker import *
 
 if __name__ == "__main__":
-    query = "What is the main idea of the paper?"
+    questions = [
+        "What is the main idea of the paper?",
+        "How does the proposed method work?",
+        "What is the most important key finding of the study?"
+    ]
     config = ConfigLoader.load_config()
     
-    invoker = LLMInvoker(query, config)
-    res = invoker.invoke()
-    print(res["result"])
+    invoker = LLMInvoker(questions, config)
+    results = invoker.invoke()
+    
+    for res in results:
+        print(f"Question: {res['query']}")
+        print(f"Answer: {res['result']}\n")
